+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   mcboost_basics_extensions: mcboost_basics_extensions.html
   mcboost_example: mcboost_example.html
-last_built: 2021-06-10T04:30Z
+last_built: 2021-06-11T04:38Z
 
