+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   mcboost_basics_extensions: mcboost_basics_extensions.html
   mcboost_example: mcboost_example.html
-last_built: 2021-06-13T04:33Z
+last_built: 2021-06-14T04:35Z
 
