### 2021-06-16 #3352 @danielskatz
👋 @pfistfl - the editors have decided this is in scope, so we will accept it for review. However, due to a current shortage of editors (that we are working to resolve), I'm going to put it on our waiting list until an appropriate editor is available.
