### 2021-07-04 #3453 @whedon
[REVIEW]: mcboost: Multi-Calibration Boosting for R
**Submitting author:** @pfistfl (<a href="http://orcid.org/0000-0001-8867-762X">Florian Pfisterer</a>)
**Repository:** <a href="https://github.com/mlr-org/mcboost" target ="_blank">https://github.com/mlr-org/mcboost</a>
**Version:** v.0.3.3
