+++ R/MCBoostSurvival.R
@@ -1,0 +1,231 @@
+
+# inherit from MCBoost object and just change private methods
+
+MCBoostSurv = R6::R6Class("MCBoostSurv",
+  inherit = MCBoost,
+  public = list(
+    times = NULL, 
+    #FIXME --- additional parameters? e.g. time buckets? 
+    initialize = function(
+      times,
+      max_iter=5,
+      alpha=1e-4,
+      eta=1,
+      partition=TRUE,
+      num_buckets=2,
+      bucket_strategy="simple",
+      rebucket=FALSE,
+      multiplicative=TRUE,
+      auditor_fitter=NULL,
+      subpops=NULL,
+      default_model_class=ConstantPredictor, #FIXME must be constant over time
+      init_predictor=NULL,
+      iter_sampling="none") {
+      
+      #FIXME does it make sense to include times here already? Preprocess? Checks? 
+      
+      self$times = times
+      
+      super$initialize(
+        max_iter,
+        alpha,
+        eta,
+        partition,
+        num_buckets,
+        bucket_strategy,
+        rebucket,
+        multiplicative,
+        auditor_fitter,
+        subpops,
+        default_model_class,
+        init_predictor,
+        iter_sampling
+      )
+    }
+  ),
+  
+  
+  
+  private = list(
+    update_probs = function(orig_preds, model, x, mask = NULL, audit = FALSE, ...) {
+      
+      deltas = numeric(length(orig_preds))
+      deltas[mask] = model$predict(x, ...)[mask]
+      
+      if (self$multiplicative) {
+        update_weights = exp(- self$eta * deltas)
+      } else {
+        update_weights = -(self$eta * deltas)
+      }
+      
+      new_preds = private$calc_new_preds(orig_preds, update_weights)
+      
+      if (audit) {
+        self$auditor_effects = c(self$auditor_effects, list(abs(deltas)))
+      }
+      return(new_preds)
+    },
+    
+    calc_new_preds = function(orig_preds, update_weights){
+      if (self$multiplicative) {
+        distlist = lapply(seq_len(length(orig_preds$ids)), function(x){
+          
+          model = orig_preds$wrappedModels(orig_preds$ids[[x]])
+          distr6::decorate(model , c("CoreStatistics", "ExoticStatistics"))
+          
+          orig_support = as.numeric(model$properties$support)
+          orig_survival = model$survival(orig_support)
+          
+          #calculate new survival probabilities
+          new_survival = clip_prob(pmax(orig_survival, 1e-4)*update_weights[x])
+          
+          # survival cannot be input, only cdf
+          new_cdf = 1- new_survival
+          distr6::WeightedDiscrete$new(x=orig_support, cdf = new_cdf)
+        })
+        
+        distr6::VectorDistribution$new(distlist = distlist)
+      } else {
+        distlist = lapply(seq_len(length(orig_preds$ids)), function(x){
+          
+          model = orig_preds$wrappedModels(orig_preds$ids[[x]])
+          distr6::decorate(model , c("CoreStatistics", "ExoticStatistics"))
+          
+          
+          orig_support = as.numeric(model$properties$support)
+          orig_survival = model$survival(orig_support)
+          
+          #calculate new survival probabilities
+          new_survival = clip_prob(orig_survival + update_weights[x])
+          
+          # survival cannot be input, only cdf
+          new_cdf = 1-new_survival
+          distr6::WeightedDiscrete$new(x=orig_support, cdf = new_cdf)
+        })
+        
+        distr6::VectorDistribution$new(distlist = distlist)
+      }
+    },
+    
+    
+    compute_residuals = function(prediction, labels) {
+      weighted_score_matrix =  private$calc_weighted_residuals(distribution = prediction, 
[…]
