+++ NEWS.md
@@ -1,5 +1,10 @@
 # mcboost (development version)
 
+# mcboost 0.3.1
+
+* Fixed a bug for additive weight updates, were updates went
+  in the wrong direction.
+
 # mcboost 0.3.0
 
 * First CRAN-ready version of the package.
