### 2021-07-14 dbc4c84
correct mistakes, change distr --> matrix
