+++ R/MCBoostSurv.R
@@ -1,6 +1,3 @@
-
-# inherit from MCBoost object and just change private methods
-
 MCBoostSurv = R6::R6Class("MCBoostSurv",
   inherit = MCBoost,
   public = list(
@@ -43,89 +40,45 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       )
     }
   ),
-  
-  
-  
   private = list(
     update_probs = function(orig_preds, model, x, mask = NULL, audit = FALSE, ...) {
-      
-      deltas = numeric(length(orig_preds))
-      deltas[mask] = model$predict(x, ...)[mask]
-      
+
       if (self$multiplicative) {
         update_weights = exp(- self$eta * deltas)
       } else {
-        update_weights = -(self$eta * deltas)
+        update_weights = (self$eta * deltas)
       }
-      
+
       new_preds = private$calc_new_preds(orig_preds, update_weights)
+
       
       if (audit) {
         self$auditor_effects = c(self$auditor_effects, list(abs(deltas)))
       }
-      return(new_preds)
+      return(clip_prob(new_preds))
     },
     
     calc_new_preds = function(orig_preds, update_weights){
       if (self$multiplicative) {
-        distlist = lapply(seq_len(length(orig_preds$ids)), function(x){
-          
-          model = orig_preds$wrappedModels(orig_preds$ids[[x]])
-          distr6::decorate(model , c("CoreStatistics", "ExoticStatistics"))
-          
-          orig_support = as.numeric(model$properties$support)
-          orig_survival = model$survival(orig_support)
-          
-          #calculate new survival probabilities
-          new_survival = clip_prob(pmax(orig_survival, 1e-4)*update_weights[x])
-          
-          # survival cannot be input, only cdf
-          new_cdf = 1- new_survival
-          distr6::WeightedDiscrete$new(x=orig_support, cdf = new_cdf)
-        })
-        
-        distr6::VectorDistribution$new(distlist = distlist)
+        # multiply weight by row wise + add noise when 0 
+        pmax(orig_preds, 1e-4) * update_weights
+        #sweep(pmax(orig_preds, 1e-4), 1, update_weights, "*"))
       } else {
-        distlist = lapply(seq_len(length(orig_preds$ids)), function(x){
-          
-          model = orig_preds$wrappedModels(orig_preds$ids[[x]])
-          distr6::decorate(model , c("CoreStatistics", "ExoticStatistics"))
-          
-          
-          orig_support = as.numeric(model$properties$support)
-          orig_survival = model$survival(orig_support)
-          
-          #calculate new survival probabilities
-          new_survival = clip_prob(orig_survival + update_weights[x])
-          
-          # survival cannot be input, only cdf
-          new_cdf = 1-new_survival
-          distr6::WeightedDiscrete$new(x=orig_support, cdf = new_cdf)
-        })
-        
-        distr6::VectorDistribution$new(distlist = distlist)
+        clip_prob(orig_preds - update_weights)
       }
     },
-    
-    
+
     compute_residuals = function(prediction, labels) {
-      weighted_score_matrix =  private$calc_weighted_residuals(distribution = prediction, 
-                                                               labels = labels)
+      weighted_score_matrix =  private$calc_weighted_residuals(prediction, labels)
       rowMeans(weighted_score_matrix)
     }, 
     
     
-    
     #FIXME should be proper & eps a hyperparameter?
-    calc_weighted_residuals = function (distribution, 
+    calc_weighted_residuals = function (prediction, 
                                         labels,
                                         proper = FALSE, 
                                         eps = 1e-4){
-      #FIXME use of internal method of mlr3proba
-      mlr3proba:::assert_surv(labels)
[…]
