### 2021-07-18 #3453 @mwt
@osorensen @pfistfl it looks all clear to me. There are automated tests which all pass and there are vignettes which compile and seem to be doing what they're supposed to.
