+++ R/MCBoostSurv.R
@@ -1,7 +1,10 @@
+# Check if several time_points? 
+
 MCBoostSurv = R6::R6Class("MCBoostSurv",
   inherit = MCBoost,
   public = list(
     time_points = NULL, 
+    num_buckets_time = NULL, 
     #FIXME --- additional parameters? e.g. time buckets? 
     initialize = function(
       time_points,
@@ -10,6 +13,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       eta=1,
       partition=TRUE,
       num_buckets=2,
+      num_buckets_time = 1, 
       bucket_strategy="simple",
       rebucket=FALSE,
       multiplicative=TRUE,
@@ -22,6 +26,8 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       #FIXME does it make sense to include time_points here already? Preprocess? Checks? 
       
       self$time_points = time_points
+
+      self$num_buckets_time = num_buckets_time
       
       super$initialize(
         max_iter,
@@ -40,8 +46,12 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       )
     }
   ),
+  #FIXME kann man das vielleicht do mit MCBoost zusammenführen ? 
   private = list(
     update_probs = function(orig_preds, model, x, mask = NULL, audit = FALSE, ...) {
+      
+      deltas = numeric(length(orig_preds))
+      deltas[mask$mask] = model$predict(x, ...)[mask$mask]
 
       if (self$multiplicative) {
         update_weights = exp(- self$eta * deltas)
@@ -70,7 +80,11 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
 
     compute_residuals = function(prediction, labels) {
       weighted_score_matrix =  private$calc_weighted_residuals(prediction, labels)
-      rowMeans(weighted_score_matrix)
+      #FIX ME IBS or BS
+      
+      #rowMeans(weighted_score_matrix)
+      
+      weighted_score_matrix
     }, 
     
     
@@ -80,15 +94,9 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
                                         proper = FALSE, 
                                         eps = 1e-4){
       
+  
       
-      if (is.null(self$time_points) || !length(self$time_points)) {
-        unique_times = unique(sort(labels[, "time"]))
-      } else {
-        #FIXME use of internal method of mlr3proba
-        unique_times = mlr3proba:::c_get_unique_times(labels[, "time"], self$time_points) 
-      }
-      
-      residuals = private$calc_residual_matrix_r(prediction, labels, unique_times)
+      residuals = private$calc_residual_matrix_r(prediction, labels, self$time_points)
       
       cens_distr = survival::survfit(survival::Surv(labels[, "time"], (1 - labels[, "status"])) ~ 1)
       cens_matrix = matrix(c(cens_distr$time, cens_distr$surv), ncol = 2)
@@ -96,31 +104,31 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       #FIXME use of internal method of mlr3proba
       
       # weight the residual matrix according to Graf et.al(1999)
-      weighted_residuals = mlr3proba:::c_weight_survival_score(residuals, 
+      weighted_residuals = mlr3proba::.c_weight_survival_score(residuals, 
                                                                labels, 
-                                                               unique_times,
+                                                               self$time_points,
                                                                cens_matrix,
                                                                proper, eps)
       
       weighted_residuals = as.data.frame(weighted_residuals)
       
-      colnames(weighted_residuals) = unique_times
+      colnames(weighted_residuals) = self$time_points
       
[…]

[…]
