### 2021-07-21 e3acc40
skip some tests on solaris
