+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mcboost
 Type: Package
 Title: Multi-Calibration Boosting
-Version: 0.3.1.9000
+Version: 0.3.1.0
 Authors@R:
     c(person(given = "Florian",
              family = "Pfisterer",
