+++ tests/testthat/test_cv_predictors.R
@@ -10,6 +10,7 @@ test_that("TreeAuditorFitters work", {
 })
 
 test_that("MCBoost multicalibrate and predict_probs - CV Predictor", {
+  skip_on_cran()
   # Sonar task
   tsk = tsk("sonar")
   data = tsk$data(cols = tsk$feature_names)
@@ -26,6 +27,7 @@ test_that("MCBoost multicalibrate and predict_probs - CV Predictor", {
 })
 
 test_that("Creating own CV Predictor works with different folds", {
+  skip_on_cran()
   # Sonar task
   tsk = tsk("sonar")
   data = tsk$data(cols = tsk$feature_names)
