+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mcboost
 Type: Package
 Title: Multi-Calibration Boosting
-Version: 0.3.2.0
+Version: 0.3.3.0
 Authors@R:
     c(person(given = "Florian",
              family = "Pfisterer",
