+++ R/helpers.R
@@ -2,15 +2,15 @@
 #' @param labels [`factor`]\cr
 #'   Factor to encode.
 #' @examples
-#'   one_hot(factor(c("a", "b", "a")))
+#' one_hot(factor(c("a", "b", "a")))
 #' @return [`integer`]\cr
 #'   Integer vector of encoded labels.
 #' @export
 one_hot = function(labels) {
   con = contrasts(labels, contrasts = FALSE)
-  mat = con[as.integer(labels),]
+  mat = con[as.integer(labels), ]
   rownames(mat) = NULL
-  if (ncol(mat) == 2L) mat = mat[,1L]
+  if (ncol(mat) == 2L) mat = mat[, 1L]
   return(mat)
 }
 
@@ -25,12 +25,13 @@ clip_prob = function(prob) {
 # Convert a X,y pair to a task
 # Required for interacting with 'mlr3'
 xy_to_task = function(x, y) {
+
   x = data.table::data.table(x)
   yname = "ytmp"
 
   # Safe yname
   i = 0
-  while(yname %in% names(x)) {
+  while (yname %in% names(x)) {
     i = i + 1
     yname = paste0("ytmp", i)
   }
@@ -49,9 +50,9 @@ xy_to_task = function(x, y) {
 #' @param learner [`mlr3::Learner`]
 #'   A trained learner used for initialization.
 #' @examples
-#'   library("mlr3")
-#'   l = lrn("classif.featureless")$train(tsk("sonar"))
-#'   mlr3_init_predictor(l)
+#' library("mlr3")
+#' l = lrn("classif.featureless")$train(tsk("sonar"))
+#' mlr3_init_predictor(l)
 #' @return [`function`]
 #' @export
 mlr3_init_predictor = function(learner) {
@@ -62,7 +63,12 @@ mlr3_init_predictor = function(learner) {
     }
   } else {
     function(data, ...) {
-      learner$predict_newdata(data)$prob[,1L]
+      learner$predict_newdata(data)$prob[, 1L]
     }
   }
+}
+
+
+even_bucket = function(pos, frac, min, max) {
+  min + pos / frac * (max - min)
 }
