+++ R/MCBoostSurv.R
@@ -1,39 +1,39 @@
-# Check if several time_points? 
+# Check if several time_points?
+# To-Do:
 
 MCBoostSurv = R6::R6Class("MCBoostSurv",
   inherit = MCBoost,
   public = list(
-    time_points = NULL, 
-    num_buckets_time = NULL, 
-    #FIXME --- additional parameters? e.g. time buckets? 
-    initialize = function(
-      time_points,
-      max_iter=5,
-      alpha=1e-4,
-      eta=1,
-      partition=TRUE,
-      num_buckets=2,
-      num_buckets_time = 1, 
-      bucket_strategy="simple",
-      rebucket=FALSE,
-      multiplicative=TRUE,
-      auditor_fitter=NULL,
-      subpops=NULL,
-      default_model_class=ConstantPredictor, #FIXME must be constant over time
-      init_predictor=NULL,
-      iter_sampling="none") {
-      
-      #FIXME does it make sense to include time_points here already? Preprocess? Checks? 
-      
-      self$time_points = time_points
+    # if time_points = NULL, use timepoints in data
+    time_points = NULL,
+    time_buckets = NULL,
+    bucket_strategies = c("even_splits", "quantiles"), # , "predictions"),#FIXME Darf ich das?
+    bucket_aggregation = NULL,
+    # FIXME --- additional parameters? e.g. time buckets?
 
-      self$num_buckets_time = num_buckets_time
-      
+    initialize = function(
+      time_points = NULL,
+      max_iter = 5,
+      alpha = 1e-4,
+      eta = 1,
+      # partition=TRUE,
+      num_buckets = 2, # probability_buckets
+      time_buckets = 1L,
+      bucket_strategy = "even_splits",
+      bucket_aggregation = NULL,
+      rebucket = FALSE,
+      multiplicative = TRUE,
+      auditor_fitter = NULL,
+      subpops = NULL,
+      default_model_class = ConstantPredictor, # FIXME must be constant over time
+      init_predictor = NULL,
+      iter_sampling = "none") {
+
       super$initialize(
         max_iter,
         alpha,
         eta,
-        partition,
+        # partition,
         num_buckets,
         bucket_strategy,
         rebucket,
@@ -44,207 +44,239 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         init_predictor,
         iter_sampling
       )
+
+      # FIXME does it make sense to include time_points here already? Preprocess? Checks?
+      self$time_points = assert_numeric(time_points,
+        sorted = TRUE, unique = TRUE,
+        min.len = 1, any.missing = FALSE,
+        lower = 0, null.ok = TRUE)
+
+      self$time_buckets = assert_int(time_buckets, lower = 1)
+      # if (self$time_buckets == 1L && self$partition) stop("If partition=TRUE, num_buckets musst be > 1!")
+      self$bucket_aggregation = assert_function(bucket_aggregation, null.ok = TRUE)
     }
   ),
-  #FIXME kann man das vielleicht do mit MCBoost zusammenführen ? 
+  # FIXME kann man das vielleicht do mit MCBoost zusammenfÃ¼hren ?
+
   private = list(
     update_probs = function(orig_preds, model, x, mask = NULL, audit = FALSE, ...) {
-      
-      deltas = numeric(length(orig_preds))
-      deltas[mask$mask] = model$predict(x, ...)[mask$mask]
+
+      dim_probs = dim(orig_preds)
+      deltas = matrix(0, nrow = dim_probs[1], ncol = dim_probs[2])
+
+      deltas <<- deltas
+      deltas_small <<- deltas[mask$n, mask$time]
+      n_1 <<- mask$n
[…]
