+++ R/MCBoost.R
@@ -25,15 +25,15 @@
 #'  }
 #'
 #' @examples
-#'   # See vignette for more examples.
-#'   # Instantiate the object
-#'   mc = MCBoost$new()
-#'   # Run multi-calibration on training dataset.
-#'   mc$multicalibrate(iris[1:100,1:4], factor(sample(c("A","B"), 100, TRUE)))
-#'   # Predict on test set
-#'   mc$predict_probs(iris[101:150,1:4])
-#'   # Get auditor effect
-#'   mc$auditor_effect(iris[101:150,1:4])
+#' # See vignette for more examples.
+#' # Instantiate the object
+#' mc = MCBoost$new()
+#' # Run multi-calibration on training dataset.
+#' mc$multicalibrate(iris[1:100, 1:4], factor(sample(c("A", "B"), 100, TRUE)))
+#' # Predict on test set
+#' mc$predict_probs(iris[101:150, 1:4])
+#' # Get auditor effect
+#' mc$auditor_effect(iris[101:150, 1:4])
 #' @export
 MCBoost = R6::R6Class("MCBoost",
   public = list(
@@ -63,7 +63,8 @@ MCBoost = R6::R6Class("MCBoost",
     #' @field partition [`logical`] \cr
     #'   True/False flag for whether to split up predictions by their "partition"
     #'   (e.g., predictions less than 0.5 and predictions greater than 0.5).
-    partition = NULL,
+    # FIXME
+    # partition = NULL,
 
     #' @field multiplicative [`logical`] \cr
     #'   Specifies the strategy for updating the weights (multiplicative weight vs additive).
@@ -97,6 +98,9 @@ MCBoost = R6::R6Class("MCBoost",
     #'   Auditor effect in each iteration.
     auditor_effects = list(),
 
+
+    bucket_strategies = c("simple"), # FIXME Darf ich das?
+
     #' @description
     #'   Initialize a multi-calibration instance.
     #'
@@ -107,6 +111,7 @@ MCBoost = R6::R6Class("MCBoost",
     #'   Accuracy parameter that determines the stopping condition. Default `1e-4`.
     #' @param eta  [`numeric`] \cr
     #'   Parameter for multiplicative weight update (step size). Default `1.0`.
+    # FIXME
     #' @param partition [`logical`] \cr
     #'   True/False flag for whether to split up predictions by their "partition"
     #'   (e.g., predictions less than 0.5 and predictions greater than 0.5).
@@ -142,72 +147,33 @@ MCBoost = R6::R6Class("MCBoost",
     #'   "bootstrap" uses a new bootstrap sample in each iteration.\cr
     #'   "none" uses the same dataset in each iteration.
     initialize = function(
-                 max_iter=5,
-                 alpha=1e-4,
-                 eta=1,
-                 partition=TRUE,
-                 num_buckets=2,
-                 bucket_strategy="simple",
-                 rebucket=FALSE,
-                 multiplicative=TRUE,
-                 auditor_fitter=NULL,
-                 subpops=NULL,
-                 default_model_class=ConstantPredictor,
-                 init_predictor=NULL,
-                 iter_sampling="none") {
+      max_iter = 5,
+      alpha = 1e-4,
+      eta = 1,
+      # FIXME
+      # partition=TRUE,
+      num_buckets = 2,
+      bucket_strategy = "simple",
+      rebucket = FALSE,
+      multiplicative = TRUE,
+      auditor_fitter = NULL,
+      subpops = NULL,
+      default_model_class = ConstantPredictor,
+      init_predictor = NULL,
+      iter_sampling = "none") {
 
-      self$max_iter = assert_int(max_iter)
-      self$alpha = assert_number(alpha)
[…]
