### 2021-07-25 a6f5589
current status
including loss and a \tau for example 75%-IBS
