+++ R/MCBoostSurv.R
@@ -9,6 +9,8 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
     time_buckets = NULL,
     bucket_strategies = c("even_splits", "quantiles"), # , "predictions"),#FIXME Darf ich das?
     bucket_aggregation = NULL,
+    time_eval = NULL, #like 75% IBS#
+    loss =  c("censored_brier", "brier"),
     # FIXME --- additional parameters? e.g. time buckets?
 
     initialize = function(
@@ -19,14 +21,17 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       # partition=TRUE,
       num_buckets = 2, # probability_buckets
       time_buckets = 1L,
+      time_eval = 1, #FIXME macht die Variable sinn? 
       bucket_strategy = "even_splits",
       bucket_aggregation = NULL,
       rebucket = FALSE,
+      eval_fulldata=FALSE,
       multiplicative = TRUE,
       auditor_fitter = NULL,
       subpops = NULL,
       default_model_class = ConstantPredictor, # FIXME must be constant over time
       init_predictor = NULL,
+      loss = "censored_brier",
       iter_sampling = "none") {
 
       super$initialize(
@@ -37,6 +42,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         num_buckets,
         bucket_strategy,
         rebucket,
+        eval_fulldata,
         multiplicative,
         auditor_fitter,
         subpops,
@@ -50,26 +56,23 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         sorted = TRUE, unique = TRUE,
         min.len = 1, any.missing = FALSE,
         lower = 0, null.ok = TRUE)
+      
+      self$time_eval = assert_double(time_eval, lower = 0.1, upper = 1, finite = TRUE, len =1 )
+
 
       self$time_buckets = assert_int(time_buckets, lower = 1)
       # if (self$time_buckets == 1L && self$partition) stop("If partition=TRUE, num_buckets musst be > 1!")
       self$bucket_aggregation = assert_function(bucket_aggregation, null.ok = TRUE)
+      self$loss = assert_choice(loss, choices = c("censored_brier", "brier"))
     }
   ),
   # FIXME kann man das vielleicht do mit MCBoost zusammenführen ?
 
   private = list(
-    update_probs = function(orig_preds, model, x, mask = NULL, audit = FALSE, ...) {
+    update_probs = function(orig_preds, model, x, mask = NULL,  update_sign = 1, audit = FALSE, ...) {
 
       dim_probs = dim(orig_preds)
       deltas = matrix(0, nrow = dim_probs[1], ncol = dim_probs[2])
-
-      deltas <<- deltas
-      deltas_small <<- deltas[mask$n, mask$time]
-      n_1 <<- mask$n
-      time_1 <<- mask$time
-      matrix_1 <<- mask$matrix
-      mask_1 = mask
 
       if (!is.null(self$bucket_aggregation)) {
         deltas[mask$n, mask$time] = (1 * (mask$matrix))
@@ -81,11 +84,13 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       deltas = deltas * auditor_predictions
 
       if (self$multiplicative) {
-        update_weights = exp(-self$eta * deltas)
-        new_preds = pmax(orig_preds, 1e-4) * update_weights
+        
+        update_weights = exp(- self$eta * update_sign * deltas)
+        # Add a small term to enable moving away from 0.
+        new_preds = update_weights * pmax(orig_preds, 1e-4)
       } else {
-        update_weights = (self$eta * deltas)
-        new_preds = clip_prob(orig_preds - update_weights)
+        update_weights = (self$eta * update_sign * deltas)
[…]
