### 2021-07-27 #3453 @OwenWard
@pfistfl I opened a pull request for two very small typos I spotted, one in an error message and one in the paper. Also the figure doesn't currently appear in the pdf of the article proof, but that should be an easy fix.
@osorensen otherwise everything looks good to me.
