### 2021-07-28 922ec79
update: add only return monotonically decreasing survival curve
