+++ R/helpers.R
@@ -72,3 +72,22 @@ mlr3_init_predictor = function(learner) {
 even_bucket = function(pos, frac, min, max) {
   min + pos / frac * (max - min)
 }
+
+
+make_decreasing_curve = function (prediction){
+  for(i in 1:nrow(prediction)){
+    if (!all(prediction[i,] == cummax(prediction[i, ]))){
+      message("Resulting curve is post-processed,
+              as it is not monotonically decreasing anymore.")
+      for(j in 2:ncol(prediction)){
+        if (prediction[i,j]>prediction[i,j-1]){
+          prediction[i,j] = prediction[i,j-1]
+        }
+      }
+    }
+  }
+  prediction
+}
+
+
+
