### 2021-07-28 3e7e3ba
Exclude vignettes on CRAN
