### 2021-07-29 fc98a02
disabler tests and examples, update paper
