+++ paper/paper.md
@@ -59,11 +59,7 @@ Users may also specify a fixed set of subgroups, instead of a learner, on which 
 Furthermore, `mcboost` includes utilities to guard against overfitting to the auditing dataset during post-processing.
 
 
-<p align="center">
-  <img src="MCBoost.png" />
-</p>
-<div align="center"> Fig 1. Conceptual illustration of Multi-Calibration Boosting with `mcboost`.</div>
-
+![Fig 1. Conceptual illustration of Multi-Calibration Boosting with `mcboost`.\label{fig:overview}](MCBoost.png)
 
 # Statement of need
 
