+++ R/AuditorFitters.R
@@ -103,6 +103,7 @@ RidgeAuditorFitter = R6::R6Class("RidgeAuditorFitter",
 #'   binary valued columns: `subpops` passed during initialization.
 #' @family AuditorFitter
 #' @examples
+#'   \dontrun{
 #'   library("data.table")
 #'   data = data.table(
 #'     "AGE_NA" = c(0, 0, 0, 0, 0),
@@ -116,6 +117,7 @@ RidgeAuditorFitter = R6::R6Class("RidgeAuditorFitter",
 #'   pops = list("AGE_NA", "AGE_0_10", "AGE_11_20", "AGE_21_31", function(x) {x[["X1" > 0.5]]})
 #'   sf = SubpopAuditorFitter$new(subpops = pops)
 #'   sf$fit(data, label - 0.5)
+#'   }
 #' @export
 SubpopAuditorFitter = R6::R6Class("SubpopAuditorFitter",
   inherit = AuditorFitter,
@@ -171,6 +173,7 @@ SubpopAuditorFitter = R6::R6Class("SubpopAuditorFitter",
 #'   binary `subgroup_masks` passed during initialization.
 #' @family AuditorFitter
 #' @examples
+#'  \dontrun{
 #'  library("data.table")
 #'  data = data.table(
 #'    "AGE_0_10" =  c(1, 1, 0, 0, 0),
@@ -185,6 +188,7 @@ SubpopAuditorFitter = R6::R6Class("SubpopAuditorFitter",
 #'    "M2" = c(1L, 0L, 0L, 0L, 1L)
 #'  )
 #'  sg = SubgroupAuditorFitter$new(masks)
+#'  }
 #' @export
 SubgroupAuditorFitter = R6::R6Class("SubgroupAuditorFitter",
   inherit = AuditorFitter,
