+++ R/PipeOpMCBoost.R
@@ -39,10 +39,10 @@
 #' Only methods inherited from [`mlr3pipelines::PipeOp`].
 #'
 #' @examples
+#' \dontrun{
 #' library(mlr3)
 #' library(mlr3pipelines)
-#' # Attention: gunion inputs have to be in the correct order for now.
-#' \dontrun{
+#' # Attention: gunion inputs have to be in the correct order for now
 #'  gr = gunion(list(
 #'    "data" = po("nop"),
 #'    "prediction" = po("learner_cv", lrn("classif.rpart"))
@@ -155,8 +155,10 @@ PipeOpMCBoost = R6Class("PipeOpMCBoost",
 #'   Note: An initial predictor can also be supplied via the `init_predictor` parameter.
 #' @return (mlr3pipelines) [`Graph`]
 #' @examples
+#'   \dontrun{
 #'   library("mlr3pipelines")
 #'   gr = ppl_mcboost()
+#'   }
 #' @export
 ppl_mcboost = function(learner = lrn("classif.featureless")) {
   mlr3misc::require_namespaces("mlr3pipelines")
