+++ cran-comments.md
@@ -1,0 +1,22 @@
+## Reason for early submission
+
+- Patch a bug in the prediction update logic
+- Fix issues raised in https://cran.r-project.org/web/checks/check_results_mcboost.html
+
+## Test Results
+
+- Issues raised by CRAN checks have been adressed
+- No new NOTEs, WARNINGs or ERRORs
+
+## R-HUB
+All r-hub checks pass without NOTEs, WARNINGs or ERRORs.
+PREPERROR on Debian Linux, R-devel, GCC likely is likely due to https://github.com/r-hub/rhub/issues/448.
+
+
+Solaris output with Vignettes:
+
+```r
+> rhub::check_on_solaris(check_args="")
+```
+
+Returns no NOTEs, WARNINGs or ERRORs
