### 2021-07-30 #3453 @osorensen
@pfistfl, I have one final comment to the article proof. On page 2, lines 52-54, I believe the last part of the sentence is not grammatically correct, particularly the word "helps".

>We hope that with mcboost, Multi-Calibration Boosting can be utilized by a wide community
53 of developers and data scientists to audit and post-process prediction models and helps to
54 promote fairness in machine learning and statistical estimation applications.

Here is a suggested rewrite, put feel free to phrase it differently:

>We hope that mcboost lets Multi-Calibration Boosting be utilized by a wide community
53 of developers and data scientists to audit and post-process prediction models, and helps to
54 promote fairness in machine learning and statistical estimation applications.
