+++ R/PipeOpMCBoostSurv.R
@@ -1,93 +1,91 @@
 PipeOpMCBoostSurv = R6Class("PipeOpMCBoostSurv",
-                        inherit = mlr3pipelines::PipeOp,
-                        public = list(
-                          
-         
-                          initialize = function(id = "mcboostsurv", param_vals = list()) {
-                            param_set = paradox::ParamSet$new(list(
-                              paradox::ParamInt$new("max_iter", lower = 0L, upper = Inf, default =5L, tags = "train"),
-                              paradox::ParamDbl$new("alpha", lower = 0, upper = 1, default = 1e-4, tags = "train"),
-                              paradox::ParamDbl$new("eta", lower = 0, upper = 1, default = 1, tags = "train"),
-                              #paradox::ParamLgl$new("partition", tags = "train", default = TRUE),
-                              paradox::ParamInt$new("num_buckets", lower = 1, upper = Inf, default = 2L, tags = "train"),
-                              paradox::ParamInt$new("time_buckets", lower = 1, upper = Inf, default = 1L, tags = "train"),
-                              paradox::ParamDbl$new("time_eval", lower = 0, upper = 1, default = 1, tags = "train"),
-                              #bucket_strategy
-                              #bucket_aggragation
-                              #eval_fulldata
-                              paradox::ParamLgl$new("rebucket", default = FALSE, tags = "train"),
-                              paradox::ParamLgl$new("multiplicative", default = TRUE, tags = "train"),
-                              paradox::ParamUty$new("auditor_fitter", default = NULL, tags = "train"),
-                              paradox::ParamUty$new("subpops", default = NULL, tags = "train"),
-                              #paradox::ParamUty$new("default_model_class", default = ConstantPredictor, tags = "train"),
-                              paradox::ParamUty$new("init_predictor", default = NULL, tags = "train")
-                            ))
-                            super$initialize(id, param_set = param_set, param_vals = param_vals, packages = character(0),
-                                             input = data.table(name = c("data", "prediction"), train = c("TaskSurv", "TaskSurv"), predict = c("TaskSurv", "TaskSurv")),
-                                             output = data.table(name = "output", train = "NULL", predict = "PredictionSurv"),
-                                             tags = "target transform")
-                          }
-                        ),
-                        private = list(
-                          .train = function(inputs) {
-                            d = inputs$data$data(cols = inputs$data$feature_names)
-                            l = inputs$data$data(cols = inputs$data$target_names)
-                            
-                            args = self$param_set$get_values(tags = "train")
-                            
-                            if (is.null(args$init_predictor)) {
-                              # Construct an initial predictor from the input model if non is provided.
-                              init_predictor = function(data, prediction) {
-                                # Prob or response prediction
-                                if (length(prediction$feature_names) > 1L) {
-                                  prds = prediction$data(cols = prediction$feature_names)
-                                  as.matrix(prds)
-                                } else {
-                                  prds = prediction$data(cols = prediction$feature_names)[[1]]
-                                  one_hot(prds)
-                                }
-                              }
-                              args$init_predictor = init_predictor
-                            }
-                            mc = invoke(MCBoostSurv$new, .args = args)
-                            mc$multicalibrate(d, l, predictor_args = inputs$prediction)
-                            self$state = list("mc" = mc)
[…]
