+++ R/Predictor.R
@@ -111,6 +111,8 @@ LearnerPredictor = R6::R6Class("LearnerPredictor",
           p = one_hot(prd$response)
         }
         return(p)
+      } else if (inherits(prd,"PredictionSurv")){
+        return(as.data.table(prd)$distr[[1]][[1]])
       }
     }
   ),
