### 2021-08-02 288af21
more efficient loops + style
