+++ R/MCBoostSurv.R
@@ -1,4 +1,4 @@
-#label must be object of Surv
+# label must be object of Surv
 
 MCBoostSurv = R6::R6Class("MCBoostSurv",
   inherit = MCBoost,
@@ -8,8 +8,8 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
     time_buckets = NULL,
     bucket_strategies = c("even_splits", "quantiles"), # , "predictions"),#FIXME Darf ich das?
     bucket_aggregation = NULL,
-    time_eval = NULL, #like 75% IBS#
-    loss =  c("censored_brier", "brier"),
+    time_eval = NULL, # like 75% IBS#
+    loss = c("censored_brier", "brier"),
     # FIXME --- additional parameters? e.g. time buckets?
 
     initialize = function(
@@ -20,11 +20,11 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       # partition=TRUE,
       num_buckets = 2, # probability_buckets
       time_buckets = 1L,
-      time_eval = 1, #FIXME macht die Variable sinn?
+      time_eval = 1, # FIXME macht die Variable sinn?
       bucket_strategy = "even_splits",
       bucket_aggregation = NULL,
       rebucket = FALSE,
-      eval_fulldata=FALSE,
+      eval_fulldata = FALSE,
       multiplicative = TRUE,
       auditor_fitter = NULL,
       subpops = NULL,
@@ -56,7 +56,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         min.len = 1, any.missing = FALSE,
         lower = 0, null.ok = TRUE)
 
-      self$time_eval = assert_double(time_eval, lower = 0.1, upper = 1, finite = TRUE, len =1 )
+      self$time_eval = assert_double(time_eval, lower = 0.1, upper = 1, finite = TRUE, len = 1)
 
 
       self$time_buckets = assert_int(time_buckets, lower = 1)
@@ -69,7 +69,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
   # FIXME kann man das vielleicht do mit MCBoost zusammenführen ?
 
   private = list(
-    update_probs = function(orig_preds, model, x, mask = NULL,  update_sign = 1, audit = FALSE, ...) {
+    update_probs = function(orig_preds, model, x, mask = NULL, update_sign = 1, audit = FALSE, ...) {
 
       dim_probs = dim(orig_preds)
       deltas = matrix(0, nrow = dim_probs[1], ncol = dim_probs[2])
@@ -85,7 +85,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
 
       if (self$multiplicative) {
 
-        update_weights = exp(- self$eta * update_sign * deltas)
+        update_weights = exp(-self$eta * update_sign * deltas)
         # Add a small term to enable moving away from 0.
         new_preds = update_weights * pmax(orig_preds, 1e-4)
       } else {
@@ -97,19 +97,18 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         auditor_effects = c(auditor_effects, list(abs(deltas)))
       }
 
-      #also correct for survival property: monotonically decreasing & between 0 and 1
+      # also correct for survival property: monotonically decreasing & between 0 and 1
       return(make_survival_curve(clip_prob(new_preds)))
     },
 
     compute_residuals = function(prediction, labels) {
-
       residuals = private$calc_residual_matrix_r(prediction, labels)
 
-      if(self$loss == "brier"){
+      if (self$loss == "brier") {
         residuals
       }
 
-      if(self$loss == "censored_brier"){
[…]
