+++ R/MCBoostSurv.R
@@ -9,18 +9,19 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
     bucket_strategies = c("even_splits", "quantiles"), # , "predictions"),#FIXME Darf ich das?
     bucket_aggregation = NULL,
     time_eval = NULL, # like 75% IBS#
-    loss = c("censored_brier", "brier"),
+    time_points_eval = NULL,
+    loss = c("censored_brier", "brier", "censored_brier_proper"),
     # FIXME --- additional parameters? e.g. time buckets?
 
     initialize = function(
-      time_points = NULL,
+      # time_points = NULL,
       max_iter = 5,
       alpha = 1e-4,
       eta = 1,
       # partition=TRUE,
       num_buckets = 2, # probability_buckets
       time_buckets = 1L,
-      time_eval = 1, # FIXME macht die Variable sinn?
+      time_eval = 1,
       bucket_strategy = "even_splits",
       bucket_aggregation = NULL,
       rebucket = FALSE,
@@ -28,7 +29,7 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       multiplicative = TRUE,
       auditor_fitter = NULL,
       subpops = NULL,
-      default_model_class = LearnerSurvKaplan, # FIXME must be constant over time
+      default_model_class = LearnerSurvKaplan,
       init_predictor = NULL,
       loss = "censored_brier",
       iter_sampling = "none") {
@@ -50,19 +51,11 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
         iter_sampling
       )
 
-      # FIXME does it make sense to include time_points here already? Preprocess? Checks?
-      self$time_points = assert_numeric(time_points,
-        sorted = TRUE, unique = TRUE,
-        min.len = 1, any.missing = FALSE,
-        lower = 0, null.ok = TRUE)
-
       self$time_eval = assert_double(time_eval, lower = 0.1, upper = 1, finite = TRUE, len = 1)
-
-
       self$time_buckets = assert_int(time_buckets, lower = 1)
       # if (self$time_buckets == 1L && self$partition) stop("If partition=TRUE, num_buckets musst be > 1!")
       self$bucket_aggregation = assert_function(bucket_aggregation, null.ok = TRUE)
-      self$loss = assert_choice(loss, choices = c("censored_brier", "brier"))
+      self$loss = assert_choice(loss, choices = c("censored_brier", "brier", "censored_brier_proper"))
 
     }
   ),
@@ -98,23 +91,22 @@ MCBoostSurv = R6::R6Class("MCBoostSurv",
       }
 
       # also correct for survival property: monotonically decreasing & between 0 and 1
-      return(make_survival_curve(clip_prob(new_preds)))
+      survival_curve = make_survival_curve(clip_prob(new_preds))
+      return(survival_curve)
     },
 
     compute_residuals = function(prediction, labels) {
-      residuals = private$calc_residual_matrix_r(prediction, labels)
-
+      residuals = private$calc_residual_matrix(prediction, labels)
       if (self$loss == "brier") {
-        return(residuals)
+        return(as.data.table(as.data.frame(residuals)))
       }
 
-      if (self$loss == "censored_brier") {
-        proper = FALSE
-        eps = 1e-4
+      if (self$loss == "censored_brier" || self$loss == "censored_brier_proper") {
 
-        cens_distr = survival::survfit(survival::Surv(labels[, "time"], (1 - labels[, "status"])) ~ 1)
+        proper = ifelse(self$loss == "censored_brier", FALSE, TRUE)
+
[…]
