+++ R/MCBoost.R
@@ -25,17 +25,17 @@
 #'  }
 #'
 #' @examples
-#'   # See vignette for more examples.
-#'   # Instantiate the object
-#'   \dontrun{
-#'   mc = MCBoost$new()
-#'   # Run multi-calibration on training dataset.
-#'   mc$multicalibrate(iris[1:100,1:4], factor(sample(c("A","B"), 100, TRUE)))
-#'   # Predict on test set
-#'   mc$predict_probs(iris[101:150,1:4])
-#'   # Get auditor effect
-#'   mc$auditor_effect(iris[101:150,1:4])
-#'   }
+#' # See vignette for more examples.
+#' # Instantiate the object
+#' \dontrun{
+#' mc = MCBoost$new()
+#' # Run multi-calibration on training dataset.
+#' mc$multicalibrate(iris[1:100, 1:4], factor(sample(c("A", "B"), 100, TRUE)))
+#' # Predict on test set
+#' mc$predict_probs(iris[101:150, 1:4])
+#' # Get auditor effect
+#' mc$auditor_effect(iris[101:150, 1:4])
+#' }
 #' @export
 MCBoost = R6::R6Class("MCBoost",
   public = list(
@@ -120,10 +120,10 @@ MCBoost = R6::R6Class("MCBoost",
     #' @param eta  [`numeric`] \cr
     #'   Parameter for multiplicative weight update (step size). Default `1.0`.
     # FIXME
-        #' @param partition [`logical`] \cr
-        #'   True/False flag for whether to split up predictions by their "partition"
-        #'   (e.g., predictions less than 0.5 and predictions greater than 0.5).
-        #'   Defaults to `TRUE` (multi-accuracy boosting).
+    #' @param partition [`logical`] \cr
+    #'   True/False flag for whether to split up predictions by their "partition"
+    #'   (e.g., predictions less than 0.5 and predictions greater than 0.5).
+    #'   Defaults to `TRUE` (multi-accuracy boosting).
     #' @param num_buckets [`integer`] \cr
     #'   The number of buckets to split into in addition to using the whole sample. Default `2L`.
     #' @param bucket_strategy  [`character`] \cr
@@ -162,21 +162,20 @@ MCBoost = R6::R6Class("MCBoost",
     #'   "none" uses the same dataset in each iteration.
 
     initialize = function(
-
-                 max_iter=5,
-                 alpha=1e-4,
-                 eta=1,
-                 # partition=TRUE,
-                 num_buckets=2,
-                 bucket_strategy="simple",
-                 rebucket=FALSE,
-                 eval_fulldata=FALSE,
-                 multiplicative=TRUE,
-                 auditor_fitter=NULL,
-                 subpops=NULL,
-                 default_model_class=ConstantPredictor,
-                 init_predictor=NULL,
-                 iter_sampling="none") {
+      max_iter = 5,
+      alpha = 1e-4,
+      eta = 1,
+      # partition=TRUE,
+      num_buckets = 2,
+      bucket_strategy = "simple",
+      rebucket = FALSE,
+      eval_fulldata = FALSE,
+      multiplicative = TRUE,
+      auditor_fitter = NULL,
+      subpops = NULL,
+      default_model_class = ConstantPredictor,
+      init_predictor = NULL,
+      iter_sampling = "none") {
 
       self$max_iter = assert_int(max_iter, lower = 0)
       self$alpha = assert_number(alpha, lower = 0)
@@ -187,7 +186,7 @@ MCBoost = R6::R6Class("MCBoost",
       self$bucket_strategy = assert_choice(bucket_strategy, choices = self$bucket_strategies)
       self$rebucket = assert_flag(rebucket)
       self$eval_fulldata = assert_flag(eval_fulldata)
[…]

[…]
