### 2021-08-03 #3453 @pfistfl
@osorensen, thanks so much for the quick & smooth progress
- tagged release: [v.0.3.3](https://github.com/mlr-org/mcboost/releases/tag/v.0.3.3)
- DOI: [10.5281/zenodo.5156518](https://zenodo.org/record/5156518)
